Initial refactor of MapView
