+++ frontend/src/redux/actions/map.js
@@ -1,0 +1,6 @@
+export const UPDATE_MAP_CENTER = "UPDATE_MAP_CENTER";
+
+export const updateMapCenter = payload => ({
+  type: UPDATE_MAP_CENTER,
+  payload
+});
