+++ frontend/src/redux/reducers/index.js
@@ -3,9 +3,12 @@ import loading from "./loading";
 import auth from "./auth";
 import resources from "./resources";
 import modal from "./modal";
+import map from "./map";
+
 export default combineReducers({
   isLoading: loading,
   auth,
   resources,
-  modal
+  modal,
+  map
 });
