+++ frontend/src/redux/selectors/map.js
@@ -1,0 +1,2 @@
+// import { createSelector } from "reselect";
+// import { resourceSelector } from "./resource";
