+++ frontend/src/redux/selectors/resource.js
@@ -1,0 +1,1 @@
+export const resourceSelector = state => state.resources;
